refactor: page 폴더 내에서 사용하는 컴포넌트로 이동
