+++ src/pages/addCard/AddCardPage.tsx
@@ -1,7 +1,7 @@
 import styles from "./AddCardPage.module.css";
-import AddCardForm from "@/AddCard/AddCardForm/components/AddCardForm";
-import AddCardPreview from "@/AddCard/AddCardPreview/components/AddCardPreview";
-import useAddCard from "@/AddCard/hooks/useAddCard";
+import AddCardForm from "@/pages/addCard/components/AddCardForm/components/AddCardForm";
+import AddCardPreview from "@/pages/addCard/components/AddCardPreview/components/AddCardPreview";
+import useAddCard from "@/pages/addCard/hooks/useAddCard";
 
 function AddCardPage() {
   const {
