+++ src/pages/addCard/components/AddCardPreview/utils/index.ts
